requirements.txt
flask
flask-cors
google-genai
requests
beautifulsoup4
python-dotenv
